billing/requirements.txt
boto3
requests
datetime
python-dateutil